refactor: extract FormValidator class
